Update index.php
Couldn't Publish the example item before FIX

Creating a new meli object in the POST code section (Publish item example) and authorizing again gave an error because we already had an ACCESS TOKEN. This gave us an empty array to '$user'. So when assigning the variables to '$_SESSION' it was set to blank. Leading us to ERROR 403. Therefore, couldn't POST the example